fix: buscar empresa no modal de adicionar usuário
